requirements.txt
bottle
Pillow
redis
pytelegrambotapi
